Removed unused methods.

+++ backend/internal/domains/auction/auction_handler.go
@@ -32,17 +32,19 @@ func NewHandler(service AuctionServiceInterface, sched scheduler.SchedulerInterf
 	}
 }
 
-// @Summary		Create Auction
-// @Description	Creates a new auction with the provided details
-// @Tags			auction
-// @Accept			json
-// @Produce		json
-// @Param			body	body		CreateAuctionDTO		true	"Auction details"
-// @Success		201		{object}	RetrieveAuctionDTO		"Created auction"
-// @Failure		400		{object}	custom_errors.HTTPError	"Bad request"
-// @Failure		401		{object}	custom_errors.HTTPError	"Unauthorized"
-// @Router			/auction [post]
-// @Security		BearerAuth
+// CreateAuction godoc
+//
+//	@Summary		Create Auction
+//	@Description	Creates a new auction with the provided details
+//	@Tags			auction
+//	@Accept			json
+//	@Produce		json
+//	@Param			body	body		CreateAuctionDTO						true	"Auction details"
+//	@Success		201		{object}	sale_offer.RetrieveDetailedSaleOfferDTO	"Created auction"
+//	@Failure		400		{object}	custom_errors.HTTPError					"Bad request"
+//	@Failure		401		{object}	custom_errors.HTTPError					"Unauthorized"
+//	@Router			/auction [post]
+//	@Security		BearerAuth
 func (h *Handler) CreateAuction(c *gin.Context) {
 	userID, err := auth.GetUserID(c)
 	if err != nil {
@@ -54,7 +56,7 @@ func (h *Handler) CreateAuction(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
 	}
-	in.UserID = (uint)(userID)
+	in.UserID = userID
 	dto, err := h.service.Create(&in)
 	if err != nil {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
@@ -105,7 +107,7 @@ func (h *Handler) DeleteAuctionByID(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
 	}
-	err = h.service.Delete(uint(id), uint(userID))
+	err = h.service.Delete(uint(id), userID)
 	if err != nil {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
@@ -120,10 +122,10 @@ func (h *Handler) DeleteAuctionByID(c *gin.Context) {
 //	@Tags			auction
 //	@Accept			json
 //	@Produce		json
-//	@Param			body	body		UpdateAuctionDTO		true	"Auction details"
-//	@Success		200		{object}	RetrieveAuctionDTO		"Updated auction"
-//	@Failure		400		{object}	custom_errors.HTTPError	"Bad request"
-//	@Failure		401		{object}	custom_errors.HTTPError	"Unauthorized"
+//	@Param			body	body		UpdateAuctionDTO						true	"Auction details"
+//	@Success		200		{object}	sale_offer.RetrieveDetailedSaleOfferDTO	"Updated auction"
+//	@Failure		400		{object}	custom_errors.HTTPError					"Bad request"
+//	@Failure		401		{object}	custom_errors.HTTPError					"Unauthorized"
 //	@Router			/auction [put]
 //	@Security		BearerAuth
 func (h *Handler) UpdateAuction(c *gin.Context) {
@@ -137,7 +139,7 @@ func (h *Handler) UpdateAuction(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
 	}
-	dto, err := h.service.Update(&auctionInput, uint(userID))
+	dto, err := h.service.Update(&auctionInput, userID)
 	if err != nil {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
